move code into routes, move test to applicable tests
