+++ app/routes.py
@@ -18,6 +18,7 @@ from dotenv import load_dotenv
 from flask import (
     Blueprint,
     Response,
+    current_app,
     flash,
     jsonify,
     make_response,
@@ -38,7 +39,6 @@ from harvester.utils.general_utils import (
     is_it_true,
     is_valid_uuid4,
     process_job_complete_percentage,
-    render_block,
 )
 
 from . import htmx
@@ -75,6 +75,27 @@ AUTH_URL = ISSUER + "/openid_connect/authorize"
 TOKEN_URL = ISSUER + "/api/openid_connect/token"
 
 STATUS_STRINGS_ENUM = {"404": "Not Found"}
+
+
+class UnsafeTemplateEnvError(RuntimeError):
+    pass
+
+
+def render_block(template_name: str, block_name: str, **context) -> Response:
+    """
+    Render a specific block from a Jinja template, while using the Flask's default environment.
+    """
+    env = current_app.jinja_env
+    if not getattr(env, "autoescape", None):
+        raise UnsafeTemplateEnvError(
+            "Jinja autoescape is disabled; enable it or use Flask's jinja_env."
+        )
+    template = env.get_template(template_name)
+
+    # Render only the named block (Jinja will still escape vars inside the block)
+    block_gen = template.blocks[block_name]
+    html = "".join(block_gen(template.new_context(context)))
+    return Response(html, mimetype="text/html; charset=utf-8")
 
 
 def login_required(f):
